pkg/core: move APIConfiguration components to call site
